Use caddy+forwardproxy instead of haproxy+tinyproxy
I wanted to see if I can manage this with forwardproxy working behind
haproxy (so that we can easily add other proxies to the same server),
but this does not seem to work, so this sets up caddy as the main web
server.

The new config has the benefit of being much simpler, since caddy can
do everything, including obtaining a TLS certificate, so we don't even
need acme.sh anymore.